Native/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart
pypdf
pillow
pytesseract